fix: (#91) Message의 폴더 위치를 변경한다.
